refactor: hidden columns into own file
